.js -> .tsx + lint
